% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessPlanR-package.R
\docType{package}
\name{businessPlanR-package}
\alias{businessPlanR}
\alias{businessPlanR-package}
\title{\packageTitle{businessPlanR}}
\description{
\packageDescription{businessPlanR}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab businessPlanR\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2024-06-20\cr
Depends: \tab R (>= 4.0.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://www.c3s.cc\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.c3s.cc}
  \item Report bugs at \url{https://github.com/C3S/businessPlanR/issues}
}

}
\author{
\packageAuthor{businessPlanR}

Maintainer: \packageMaintainer{businessPlanR}
}
