% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_businessplan.R
\name{pdf_businessplan}
\alias{pdf_businessplan}
\title{R Markdown output format based on \code{bookdown::pdf_document2}}
\usage{
pdf_businessplan(
  base_format = bookdown::pdf_document2,
  link_dirs = c(),
  copy_files = c(),
  makeindex = FALSE,
  makeglossary = FALSE,
  pdflatex_runs = "auto",
  keep_tex = TRUE,
  ...
)
}
\arguments{
\item{base_format}{A function, the base format to extend.}

\item{link_dirs}{Vector of directories to symlink to the build directory,
      in absolute paths.}

\item{copy_files}{Vector of files to copy to the build directory, in absolute paths.}

\item{makeindex}{Logical, whether to invoke makeindex after pdflatex.}

\item{makeglossary}{Logical, whether to invoke makeglossary after pdflatex.}

\item{pdflatex_runs}{Either an integer number,
      the number of times pdflatex (and makeindex/makeglossary, respectively)
should run to render the document.}

\item{...}{Further options to pass to \code{base_format}.}
}
\value{
An R Markdown output format object to be passed to \code{rmarkdown::render()}.
}
\description{
This output format definition extends \code{bookdown::pdf_document2} in two ways.
Firstly, it allows to manually define directories and files that should be
copied or linked to the build directory. Secondly, you can manually set
the number of \code{pdflatex} runs.
}
\details{
While working on a more complex business plan including a glossary and
an index with keywords, we encountered the problem that \code{rmarkdown::render()}
did one run of \code{pdflatex} too few to produce a PDF file with correct page numbers.
As a workaround, we had it to render the LaTeX file only, copied missing
files manually and ran \code{pdflatex} as often as necessary.

We generalized this workaround into this output format definition for
\code{rmarkdown::render()}.
}
